safesearchesx/requirements.txt
Flask==2.3.2
python-dotenv==1.0.0
Authlib==1.2.0
google-api-python-client==2.86.0
requests==2.29.0
pyOpenSSL==23.1.1
urllib3==1.26.15
gunicorn==20.1.0
whois